delete endpoint authentication

+++ TodoApp/routers/todos.py
@@ -84,12 +84,15 @@ async def update_todo(db:db_dependency,
 
 
 @router.delete("/todo/delete/{todo_id}", status_code= status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,
+async def delete_todo(db:db_dependency,user: user_dependency,
                       todo_id : int = Path(gt = 0)):
-     
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+        
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is  None:
         raise HTTPException(status_code = 404 , detail="Todo not found")
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
